server/pyproject.toml
[tool.poetry]
name = "code-puzzles"
version = "0.1.0"
description = ""
authors = ["Serhii Lesyk <97725518+serjalexle@users.noreply.github.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
pydantic = {extras = ["email"], version = "^2.9.2"}
apscheduler = "^3.10.4"
aiohttp = "^3.10.10"
motor = "^3.6.0"
beanie = "^1.27.0"
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"
python-multipart = "^0.0.20"
python-socketio = {extras = ["asyncio"], version = "^5.13.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


